Add autogenerator value saving
